R18 audit: 5 fixes — heartbeat await, ROLLBACK on BEGIN fail, connect_direct guard, pool_size u64, before_acquire config
F2: begin_direct now sends best-effort ROLLBACK before dropping a
connection when BEGIN fails. In pooler environments, this ensures the
server-side connection is cleanly released rather than relying solely
on TCP close detection.

F3: shutdown() now stores the heartbeat JoinHandle and awaits its
graceful exit (15s timeout) after cancelling the CancellationToken.
Previously the JoinHandle was dropped, leaving the heartbeat task
running until the tokio runtime shut down — potentially leaking the
heartbeat connection for minutes.

F14: connect_direct's debug_assert for zero connect_timeout is now a
release-safe defensive check that returns PgStoreError::Other with a
clear message. The debug_assert is kept as a secondary panic in debug
builds.

F5: pool_size() returns (u32, u64) instead of (u32, u32). The idle
count came from pool.num_idle() (usize) truncated to u32; now cast
directly to u64 to eliminate theoretical truncation on platforms
where usize > u32.

F19: before_acquire's 60s hardcoded ping threshold now uses
tune.keepalive_idle (default 60s). If a user customizes
PG_TUNED_KEEPALIVE_IDLE, the pool ping threshold stays aligned with
the heartbeat interval.

All 91 tests pass, clippy zero warnings.

+++ src/pg_builder.rs
@@ -105,7 +105,7 @@ impl TransactionBuilder for PgStore {
                 if metric == "pg_pool_size" {
                     Some(size as u64)
                 } else {
-                    Some(idle as u64)
+                    Some(idle) // already u64 from pool_size()
                 }
             }
             "pg_pool_max" => Some(self.pool_max() as u64),
